Queues Update
Queue cmd now works as expected. Checks if file exists. If not it creates a new file with the new queue. If file does exist and there are no queues, it creates one. If the file does exist and the queue also exists then it will notify the player that the queue exists and to use /join cmd.